Final Readme.md

+++ src/main/java/com/example/freview/repositories/SerieRepository.java
@@ -3,6 +3,6 @@ package com.example.freview.repositories;
 import com.example.freview.models.Serie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SerieRepository extends JpaRepository<Serie,Integer> {
+public interface SerieRepository extends JpaRepository<Serie,Long> {
 
 }
